requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.8
autogen==0.7.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
fast-depends==2.4.12
fastapi==0.115.8
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.3
openai==1.63.2
packaging==24.2
pyautogen==0.7.4
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.45.3
termcolor==2.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
# psycopg2
psycopg2-binary
passlib[bcrypt]
sqlalchemy
python-jose
pydantic-settings


